+++ awslimitchecker/services/ec2.py
@@ -243,6 +243,8 @@ class _Ec2Service(_AwsService):
             'i2.8xlarge': (2, 20, 0),
             'd2.4xlarge': (10, 20, 5),
             'd2.8xlarge': (5, 20, 5),
+            'm4.4xlarge': (10, 20, 5),
+            'm4.10xlarge': (5, 20, 5)
         }
         limits = {}
         for i_type in self._instance_types():
@@ -413,6 +415,7 @@ class _Ec2Service(_AwsService):
         :rtype: list
         """
         GENERAL_TYPES = [
+            't2.nano',
             't2.micro',
             't2.small',
             't2.medium',
@@ -425,7 +428,7 @@ class _Ec2Service(_AwsService):
             'm4.xlarge',
             'm4.2xlarge',
             'm4.4xlarge',
-            'm4.8xlarge',
+            'm4.10xlarge'
         ]
 
         PREV_GENERAL_TYPES = [